murdock: nordic_softdevice_ble: softdevice.hex is a test input
When running tests using 'nordic_softdevice_ble', the 'softdevice.hex'
file must also be taken into account for the test hashing and be
uploaded to the separated murdock testing boards.

When listed as dependency from `test-input-hash` the file must have a
target, which he has not. The file is implicely created when compiling
`ELFFILE` so depend as order-only from it.
In practice `BASELIBS` could be enough or even `pkg-build-softdevice`
but I do not want to be depend that much on internals there.