sys/shell: simplify array traversal code.
The code for traversing arrays of shell commands (used to print help messages
and to search for commmand handlers) was needlessly complex.

The new code is both cleaner and results in smaller binaries (at least for
samr21). A substantial part of the improvement in size is due to the following
change

```
-    printf("%-20s %s\n", "Command", "Description");
-    puts("---------------------------------------");
+    puts("Command              Description\n---------------------------------------");
```

The way the code is written now means that ifdefs in the middle of the code
can be eliminated, further enhancing clarity.

Size before:

```
   text	   data	    bss	    dec	    hex	filename
  10176	    140	   2612	  12928	   3280	tests/shell/bin/samr21-xpro/tests_shell.elf
```

After:

```
   text	   data	    bss	    dec	    hex	filename
  10164	    140	   2612	  12916	   3274	tests/shell/bin/samr21-xpro/tests_shell.elf
```