tests/shell: fix false negative in the Ctrl-C test.
The test for the line cancellation (ctrl-c) functionality was unable to
detect error because of the way pexpect matches output.

While working on the long line shell bug, a regression was about to be
introduced because of this. This commit fixes the test by directly reading from
the child process and expects an exact response.

To make the test more robust, the newline handling was changed in native,
where an extra empty line was being sent with each character.