sys: Revert addition of irq_handler (#10555)
The API for dispatching work to one or more dedicated worker threads is not
yet ready to be exposed to a broader public in a release, as the API is
currently under heavy discussion. There for it's addition is reverted in the
release 2019.10 release branch in the hope it is ready for a broader audience
in the next release.

The usual deprecation process is skipped in this revert, as the API was merged
with PR #10555 on 13th of August 2019; and therefore was not part of any release
yet.