sys/luid: apply lastused to every byte
The most common use case for luid_get() is go generate a hardware
address.
For this, the lowest two bits of the first byte will be set to a defined
value, so the resulting ID is only unique for every 3rd invocation.

To fix this, apply the lastused xor to every byte.