Added more FAQ entries to emcute example
I worked with the emcute example and had the added problems. I found the solutions in the notes section of the asymcute readme and I think they are needed here too, to be easily accessible for beginners. Especially the the problem with the link local addresses was hard to debug for a RIOT-OS/MQTT-SN beginner.