boards/atmega328p: Fixed xtimer frequency
As the ATmega328p on a breadboard can have various frequencies depending on
the fuse settings and whether and which crystal is connected, a fixed xtimer
frequency that can be configured (with the available prescalers) is not
possible. Therefore, the fixed frequency is replaced by a frequency depending
on the CPU clock speed. On 16 MHz, XTIMER_HZ will yield 250000UL, the same value
used on the Arduino UNO.