makefiles/color: fix build on NixOS
On NixOS, there is no `/bin/echo`, instead
`which echo` will return

    /home/benpicco/.nix-profile/bin/echo

so hard-coding the path will break here.

However, the -e command seems to be well supported by the build-in
even, so there is no need to encode an absolute path.

(Should probably be tested on some more systems, only tried Ubuntu and NixOS.)