at86rf215: allow board-specific trim value
The internal oscillator can be tuned by adding up to 16*0.3pF to the
RC circuit.
To do so connect a frequency counter to the chip's clock output
(and set AT86RF215_USE_CLOCK_OUTPUT to 1), then tune the value until the
measured frequency is closest to 26 MHz.

This should improve the radio characteristics of the transceiver.