pkg/semtech-loramac: TimerInit(): use static struct initializer
Previously, TimerInit() explicitly set some values of the xtimer truct
to "0". This commit changes it to use a static struct initializer
`(xtimer_t){0}`. The advantage is that now this is independent of the
actual xtimer structure.

Now this also resets the remaining xtimer fields, but they're properly
re-populated elsewhere (by using xtimer_set_msg()).