cpu/atmega_common: Implement RTC based on RTT
This implements a basic Real Time Clock based on the Real Time Timer
peripheral.
As the timer is too fast and wraps around after just 24 bits, it is
not used directly. Instead the RTT is responsible for providing a 1Hz
tick by generating an alarm every second.

The current time data is kept in the `.noinit` section, so it will survive
a reboot, but the clock will not be updated while the bootloader runs, so
expect inaccuracies.