cpu/samd5x: allow to use XOSC as clock source
Allow to run the main clock and all peripheral clocks off XOSC.
This is necessary if we want to use the buck voltage regulator.

We can always use the buck regulator in standby mode (and below)
when the fast internal oscillators are switched off.

Switch back to LDO after wake-up when one of the fast oscillators is used.