boards/same54-xpro: provide config to run off XOSC
The board provides a 12 MHz oscillator that can be used to source
the CPU and peripheral clocks.
When doing so we don't have to rely on the internal fast oscillators.
This allows us to enable the buck regulator for maximum power savings.

With this same54-xpro reaches 0.8 mA in IDLE.