cpu/sam0_common: derive ROM_LEN & RAM_LEN from part number
The ROM size is encoded in the part number of the Atmel SAM chips.
RAM size is not encoded directly, but it's always a fixed fraction
of the ROM size.

On SAM D2x/SAML2x it's 1/8th of the ROM.
On SAM L1x it's 1/4th of the ROM.
On SAMD5x RAM is at least 128k and increments by 64k with every ROM
increment.