tools/radvd: Add script to run radvd on a tun interface
This adds a small wrapper script to configure and start radvd on a
tun interface.
This allows the use of router advertisements and global adresses with
the `native` target.

Usage:

 - first create the tap interface as usual

    sudo dist/tools/tapsetup/tapsetup

 - now run radvd on the new tapbr0 interface

    sudo dist/tools/radvd/radvd.sh -c tapbr0 2001:db8::/64

 - Now run the `gnrc_networking` example on native:

    make -C examples/gnrc_networking all term

You should now see that the `native` node has received a global address
in `ifconfig`.
You should be able to reach this address from your host computer.

It may take very long for the native node to obtain the address.
If you are observing this, try turning off router advertisements of the
native node on the upstream interface by running

    ifconfig 6 -rtr_adv

Alternatively change `USEMODULE += gnrc_ipv6_router_default` to
`USEMODULE += gnrc_ipv6_default` in the project's `Makefile`.