sys/shell: native: stop RIOT when the shell reads EOF
Currently when the shell receives EOF on `native`, sending EOF
(ctrl + D) closes the stdin file descriptor, leading to all
consecutive reads to also return EOF.

The result is that `shell_run_forever()` will re-start the shell
forever in a loop, filling up the terminal with

    > > > > > > > > > > > > > > > > > > > > > > …

until the user hits ctrl + C.

This is annoying.

Instead, cleanly shutdown RIOT when receiving EOF on native, which
should match the expected behaviour of most users.